#309 Support importing project from RMG via app clip
