+++ package.json
@@ -10,7 +10,7 @@
         "@chakra-ui/react": "^2.4.3",
         "@emotion/react": "^11.10.5",
         "@emotion/styled": "^11.10.5",
-        "@railmapgen/rmg-components": "^7.1.2",
+        "@railmapgen/rmg-components": "^7.1.4",
         "@railmapgen/rmg-runtime": "^5.0.1",
         "@railmapgen/rmg-palette-resources": "^0.7.11",
         "@railmapgen/rmg-translate": "^2.0.1",
