+++ src/components/page-header/open-actions.tsx
@@ -2,27 +2,19 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { MdInsertDriveFile, MdNoteAdd, MdUpload } from 'react-icons/md';
 import { Badge, IconButton, Menu, MenuButton, MenuItem, MenuList } from '@chakra-ui/react';
-import { nanoid } from 'nanoid';
 import { useRootDispatch } from '../../redux';
 import {
     ParamState,
-    setSvgViewBoxZoom,
-    setSvgViewBoxMin,
     saveGraph,
     setFullState,
+    setSvgViewBoxMin,
+    setSvgViewBoxZoom,
 } from '../../redux/param/param-slice';
-import { clearSelected, setRefreshNodes, setRefreshEdges, setGlobalAlert } from '../../redux/runtime/runtime-slice';
-import { LinePathType, LineStyleType } from '../../constants/lines';
-import { StationAttributes, StationType } from '../../constants/stations';
-import { InterchangeInfo } from '../panels/details/interchange-field';
-import { linePaths } from '../svgs/lines/lines';
-import { ShmetroBasic2020StationAttributes } from '../svgs/stations/shmetro-basic-2020';
-import { GzmtrBasicStationAttributes } from '../svgs/stations/gzmtr-basic';
-import { GzmtrIntStationAttributes } from '../svgs/stations/gzmtr-int';
-import { MTRStationAttributes } from '../svgs/stations/mtr';
-import stations from '../svgs/stations/stations';
+import { clearSelected, setGlobalAlert, setRefreshEdges, setRefreshNodes } from '../../redux/runtime/runtime-slice';
 import { RMPSave, upgrade } from '../../util/save';
 import { GalleryModal } from './gallery-modal';
+import { parseRmgParam } from '../../util/rmg-param-parser';
+import RmgParamAppClip from './rmg-param-app-clip';
 
 export default function OpenActions() {
     const { t } = useTranslation();
@@ -33,6 +25,7 @@ export default function OpenActions() {
     const fileInputRef = React.useRef<HTMLInputElement | null>(null);
 
     const [isGalleryModalOpen, setIsGalleryModalOpen] = React.useState(false);
+    const [isRmgParamAppClipOpen, setIsRmgParamAppClipOpen] = React.useState(false);
 
     const refreshAndSave = React.useCallback(() => {
         dispatch(setRefreshNodes());
@@ -58,146 +51,7 @@ export default function OpenActions() {
         } else {
             try {
                 const paramStr = await readFileAsText(file);
-                const param = JSON.parse(paramStr);
-                const stnList = param.stn_list;
-                const theme = param.theme;
-
-                // generate stn id
-                const stnIdMap = Object.fromEntries(
-                    Object.keys(stnList)
-                        .filter(id => !['linestart', 'lineend'].includes(id))
-                        .map(id => [id, `stn_${nanoid(10)}`])
-                );
-                // update stnIdMap if stations exist in the graph
-                Object.entries(stnList)
-                    .filter(([id, _]) => !['linestart', 'lineend'].includes(id))
-                    .forEach(([id, stnInfo]) => {
-                        const nodes = graph.current.filterNodes(
-                            (node, attr) =>
-                                // @ts-expect-error
-                                Object.values(StationType).includes(attr.type) &&
-                                // @ts-expect-error
-                                (attr[attr.type] as StationAttributes).names[0] === stnInfo.name[0]
-                        );
-                        if (nodes.length !== 0) stnIdMap[id] = nodes[0];
-                    });
-
-                // only import stations that don't appear in the graph
-                Object.entries(stnList)
-                    .filter(([id, _]) => !['linestart', 'lineend'].includes(id))
-                    .filter(
-                        ([id, stnInfo]) =>
-                            graph.current.filterNodes(
-                                (node, attr) =>
-                                    // @ts-expect-error
-                                    Object.values(StationType).includes(attr.type) &&
-                                    // @ts-expect-error
-                                    (attr[attr.type] as StationAttributes).names[0] === stnInfo.name[0]
-                            ).length === 0
-                    )
-                    .forEach(([id, stnInfo], i) => {
-                        // determine station type
-                        let type: StationType = StationType.ShmetroBasic;
-                        if (param.style === 'shmetro') {
-                            if ((stnInfo as any).transfer.info.flat().length > 0) type = StationType.ShmetroInt;
-                            else if (param.info_panel_type === 'sh2020') type = StationType.ShmetroBasic2020;
-                            else type = StationType.ShmetroBasic;
-                        } else if (param.style === 'gzmtr') {
-                            if ((stnInfo as any).transfer.info.flat().length > 0) type = StationType.GzmtrInt;
-                            else type = StationType.GzmtrBasic;
-                        } else if (param.style === 'mtr') {
-                            type = StationType.MTR;
-                        }
-
-                        // read default attrs
-                        const attr = {
-                            // deep copy to prevent mutual reference
-                            ...JSON.parse(JSON.stringify(stations[type].defaultAttrs)),
-                            names: (stnInfo as any).name,
-                        };
-
-                        // add style specific attrs from RMG save
-                        if (type === StationType.ShmetroBasic2020)
-                            (attr as ShmetroBasic2020StationAttributes).color = theme;
-                        else if (type === StationType.GzmtrBasic) {
-                            (attr as GzmtrBasicStationAttributes).color = param.theme;
-                            (attr as GzmtrBasicStationAttributes).lineCode = param.line_num;
-                            (attr as GzmtrBasicStationAttributes).stationCode = (stnInfo as any).num;
-                        } else if (type === StationType.GzmtrInt) {
-                            const transfer = JSON.parse(
-                                JSON.stringify((stnInfo as any).transfer.info)
-                            ) as InterchangeInfo[][];
-                            // override line code and station code to default as they are not provided in RMG save
-                            transfer.forEach(lv1 =>
-                                lv1.forEach(transferInfo => {
-                                    transferInfo[4] = '1';
-                                    transferInfo[5] = '01';
-                                })
-                            );
-                            // add current line and station code to transfer[0][0]
-                            transfer[0].unshift([
-                                ...param.theme,
-                                param.line_num,
-                                (stnInfo as any).num,
-                            ] as unknown as InterchangeInfo);
-                            (attr as GzmtrIntStationAttributes).transfer = transfer;
-                        } else if (type === StationType.MTR) {
-                            let transfer = JSON.parse(
-                                JSON.stringify((stnInfo as any).transfer.info)
-                            ) as InterchangeInfo[][];
-                            // drop out of station transfer as they should be placed in another station
-                            transfer = [transfer[0]];
-                            // override line code and station code to empty as they are useless in MTR station
-                            transfer.forEach(level =>
-                                level.forEach(transferInfo => {
-                                    transferInfo[4] = '';
-                                    transferInfo[5] = '';
-                                })
-                            );
-                            if (transfer.flat().length > 0) {
-                                // add current theme to transfer[0][0] as MTR display all transfers including the current line
-                                transfer[0].unshift([...param.theme, '', ''] as unknown as InterchangeInfo);
-                            }
-                            (attr as MTRStationAttributes).transfer = transfer;
-                        }
-
-                        graph.current.addNode(stnIdMap[id], {
-                            visible: true,
-                            zIndex: 0,
-                            x: 100 + i * 50,
-                            y: 1000,
-                            type,
-                            [type]: attr,
-                        });
-                    });
-
-                // import lines
-                Object.entries(stnList)
-                    .filter(([id, _]) => !['linestart', 'lineend'].includes(id))
-                    .forEach(([id, stnInfo]) => {
-                        (stnInfo as any).children
-                            .filter((child: string) => !['linestart', 'lineend'].includes(child))
-                            .forEach((child: string) => {
-                                graph.current.addDirectedEdgeWithKey(
-                                    `line_${nanoid(10)}`,
-                                    stnIdMap[id],
-                                    stnIdMap[child],
-                                    {
-                                        visible: true,
-                                        zIndex: 0,
-                                        type: LinePathType.Diagonal,
-                                        // deep copy to prevent mutual reference
-                                        [LinePathType.Diagonal]: JSON.parse(
-                                            JSON.stringify(linePaths[LinePathType.Diagonal].defaultAttrs)
-                                        ),
-                                        style: LineStyleType.SingleColor,
-                                        [LineStyleType.SingleColor]: { color: theme },
-                                        reconcileId: '',
-                                    }
-                                );
-                            });
-                    });
-
+                parseRmgParam(graph.current, JSON.parse(paramStr));
                 refreshAndSave();
             } catch (err) {
                 dispatch(setGlobalAlert({ status: 'error', message: t('OpenActions.unknownError') }));
@@ -210,6 +64,18 @@ export default function OpenActions() {
 
         // clear field for next upload
         event.target.value = '';
+    };
+
+    const handleImportRMGProject = (param: Record<string, any>) => {
+        try {
+            parseRmgParam(graph.current, param);
+            refreshAndSave();
+        } catch (err) {
+            dispatch(setGlobalAlert({ status: 'error', message: t('OpenActions.unknownError') }));
+            console.error('OpenActions.handleUploadRMG():: Unknown error occurred while parsing the RMG project', err);
+        } finally {
+            setIsRmgParamAppClipOpen(false);
+        }
     };
 
     const handleUpload = async (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -292,6 +158,10 @@ export default function OpenActions() {
                     {t('header.open.configRMG')}
                 </MenuItem>
 
+                <MenuItem icon={<MdInsertDriveFile />} onClick={() => setIsRmgParamAppClipOpen(true)}>
+                    {t('header.open.projectRMG')}
+                </MenuItem>
+
                 <MenuItem icon={<MdInsertDriveFile />} onClick={() => setIsGalleryModalOpen(true)}>
                     {t('header.open.gallery')}
                     <Badge ml="1" colorScheme="green">
@@ -304,6 +174,12 @@ export default function OpenActions() {
                     onClose={() => setIsGalleryModalOpen(false)}
                 />
             </MenuList>
+
+            <RmgParamAppClip
+                isOpen={isRmgParamAppClipOpen}
+                onClose={() => setIsRmgParamAppClipOpen(false)}
+                onImport={handleImportRMGProject}
+            />
         </Menu>
     );
 }
