+++ src/i18n/translations/zh-Hant.json
@@ -256,9 +256,10 @@
         "popoverBody": "我們正在測試最新的RMP。如果妳有任何建議，歡迎在 https://github.com/railmapgen/rmp/issues 上提出。",
         "open": {
             "new": "新項目",
-            "config": "導入項目",
-            "configRMG": "從RMG存檔中導入",
-            "gallery": "從畫廊中導入",
+            "config": "讀入項目",
+            "configRMG": "從RMG存檔中讀入",
+            "projectRMG": "從RMG專案中讀入",
+            "gallery": "從畫廊中讀入",
             "shanghai": "上海",
             "guangzhou": "廣州",
             "beijing": "北京",
